+++ src/graph/ops/fork_op.rs
@@ -65,6 +65,10 @@ impl<T: Trace> GraphNode<T> for ForkOp<T> {
         self.graph.finish()
     }
 
+    fn finish_existing(&self) -> Result<()> {
+        self.graph.finish_existing()
+    }
+
     fn name(&self) -> &'static str {
         Self::NAME
     }