+++ src/graph/ops/time_op.rs
@@ -66,6 +66,10 @@ impl<T: Trace> GraphNode<T> for TimeOp<T> {
         self.graph.finish()
     }
 
+    fn finish_existing(&self) -> Result<()> {
+        self.graph.finish_existing()
+    }
+
     fn name(&self) -> &'static str {
         Self::NAME
     }