+++ src/graph/ops/while_op.rs
@@ -143,6 +143,10 @@ impl<T: Trace> GraphNode<T> for WhileOp<T> {
         self.graph.finish()
     }
 
+    fn finish_existing(&self) -> Result<()> {
+        self.graph.finish_existing()
+    }
+
     fn name(&self) -> &'static str {
         Self::NAME
     }